pyproject.toml
[tool.poetry]
name = "agent-discussion"
version = "0.1.0"
description = "Multi-agent discussion application powered by Claude"
authors = []
packages = [{ include = "agent_discussion" }]

[tool.poetry.dependencies]
python = "^3.13"
anthropic = "0.49.0"
streamlit = "1.43.2"
pyyaml = "6.0.2"
pymupdf = "1.25.3"
python-docx = "1.1.2"

[tool.poetry.group.dev.dependencies]
pytest = "8.3.5"
pytest-asyncio = "0.25.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
